+++ examples/hello/project.nix
@@ -1,18 +1,12 @@
-let
-  nedryland = import ../../default.nix;
-  project = nedryland.mkProject {
-    name = "hello-example";
-    configFile = ./config.toml;
-  };
-in
-project.mkGrid {
+(import ../../default.nix).mkProject {
+  name = "hello-example";
+
   # The keys in the components set are used to depend on or reference components
   # they declare their actual name in their respective nix files
-  components = rec {
+  components = { callFile }: {
     # Library for advanced math functions
-    numpyWrapper = project.declareComponent ./utils/numpy-wrapper/numpy-wrapper.nix { };
+    numpyWrapper = callFile ./utils/numpy-wrapper/numpy-wrapper.nix { };
     # Client that uses the above library to print a message
-    pythonHello = project.declareComponent ./clients/hello/hello.nix { numpyWrapper = numpyWrapper.package; };
+    pythonHello = callFile ./clients/hello/hello.nix { };
   };
-  deploy = { };
 }