+++ examples/protobuf/default.nix
@@ -1,1 +1,1 @@
-(import ./project.nix).grid
+(import ./project.nix).matrix