+++ examples/protobuf/project.nix
@@ -1,30 +1,17 @@
 # Example project illustrating the protobuf generation functionality in Nedryland
-let
-  nedryland = import ../../default.nix;
-
-  project = nedryland.mkProject {
-    name = "protobuf-example";
-    configFile = ./config.toml;
-  };
-
-in
-project.mkGrid {
-  components = rec {
+(import ../../default.nix).mkProject {
+  name = "protobuf-example";
+  components = { callFile }: {
     # expose the compiler here as well so it can be accessed
-    compiler = project.declareComponent ./compiler.nix { };
+    compiler = callFile ./compiler.nix { };
 
     # create a proto module that depends on another one
-    baseProtocols = project.declareComponent ./protocols/base/protocols.nix { };
-    protocols = project.declareComponent ./protocols/ext/protocols.nix {
-      baseProtos = baseProtocols;
-    };
+    baseProtocols = callFile ./protocols/base/protocols.nix { };
+    protocols = callFile ./protocols/ext/protocols.nix { };
 
     client = {
-      rust = project.declareComponent ./clients/rust/client.nix { inherit protocols baseProtocols; };
-      python = project.declareComponent ./clients/python/client.nix { inherit protocols baseProtocols; };
+      rust = callFile ./clients/rust/client.nix { };
+      python = callFile ./clients/python/client.nix { };
     };
   };
-
-
-  deploy = { };
 }