+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deploying from inside nix has been removed. Instead we create a deployment script which is called outside nix.
 
 ### Changed
+- mkGrid does not exist anymore and instead a list of components is sent to mkProject and `matrix` is a 
+  property on the project set, together with `shells`, `name`, `lib` etc. All attrs sent to `mkProject` also
+  supports functions that takes a single set with members from `base`.
 - Renamed `test` command to `check` due to bash if statement conflict (if statements actually run a function called test).
 - Use niv for handling nix dependencies.
 - Python overlays now work for both Python 3.7 and 3.8. Furthermore it is exposed both on the